chore: Organizes the files of the ReportModal feature
