+++ superset-frontend/src/dashboard/containers/DashboardHeader.jsx
@@ -57,7 +57,6 @@ import {
 
 import { logEvent } from 'src/logger/actions';
 import { DASHBOARD_HEADER_ID } from 'src/dashboard/util/constants';
-import { fetchUISpecificReport } from 'src/reports/actions/reports';
 
 function mapStateToProps({
   dashboardLayout: undoableLayout,
@@ -130,7 +129,6 @@ function mapDispatchToProps(dispatch) {
       dashboardInfoChanged,
       dashboardTitleChanged,
       updateDataMask,
-      fetchUISpecificReport,
     },
     dispatch,
   );
